+++ src/composables/index.ts
@@ -1,3 +1,4 @@
+export { useExternalTrigger } from './useExternalTrigger'
 export { useTooltipEvents } from './useTooltipEvents'
 export type { TooltipEventHandlers, TriggerType } from './useTooltipEvents'
 export { useTooltipPosition } from './useTooltipPosition'